# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Hypermedia project by Roccamena, Romanò, Buttolo"
  version: "1.0.0"
  title: "Child4Help"
  termsOfService: "http://swagger.io/terms/"
host: "localhost:8080"
basePath: "/api"
...
schemes:
- "https"
- "http"
paths:
...
  /services/items:
    get:
      tags:
      - "service"
      summary: "Get all services items"
      operationId: "getServicesItemsAll"
      produces:
      - "application/json"
      parameters:
      - name: "page"
        in: "query"
        description: "Services page number to get"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ServiceItem"
      x-swagger-router-controller: "Service"
  /service/{id_service}:
    get:
      tags:
      - "service"
      summary: "Get a service by ID"
      operationId: "getServiceByID"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id_service"
        in: "path"
        description: "ID of service to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Service"
        "400":
          description: "Invalid ID"
        "404":
          description: "Service not found"
      x-swagger-router-controller: "Service"
    delete:
      tags:
      - "service"
      summary: "Delete a service by ID"
      operationId: "deleteService"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id_service"
        in: "path"
        description: "ID of service to be deleted"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid ID"
        "404":
          description: "Service not found"
      x-swagger-router-controller: "Service"
  /service/{id_service}/page1:
    get:
      tags:
      - "service"
      summary: "Get a service page 1 data by ID"
      operationId: "getServicePage1ByID"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id_service"
        in: "path"
        description: "ID of service to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
          schema:
            $ref: "#/definitions/ServicePage1"
        "400":
          description: "Invalid ID"
        "404":
          description: "Service not found"
      x-swagger-router-controller: "Service"
  /service/{id_service}/page2:
    get:
      tags:
      - "service"
      summary: "Get a service page 2 data by ID"
      operationId: "getServicePage2ByID"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id_service"
        in: "path"
        description: "ID of service to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
          schema:
            $ref: "#/definitions/ServicePage2"
        "400":
          description: "Invalid ID"
        "404":
          description: "Service not found"
      x-swagger-router-controller: "Service"
  /service:
    post:
      tags:
      - "service"
      summary: "Add a new service"
      operationId: "addService"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - in: "body"
        name: "service"
        description: "Service object to be added"
        required: true
        schema:
          $ref: "#/definitions/Service"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid input"
      x-swagger-router-controller: "Service"
    put:
      tags:
      - "service"
      summary: "Update a service"
      operationId: "updateService"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - in: "body"
        name: "service"
        description: "Service object to be updated"
        required: true
        schema:
          $ref: "#/definitions/Service"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid input"
        "404":
          description: "Service not found"
      x-swagger-router-controller: "Service"
  /service/{id_service}/related:
    get:
      tags:
      - "service"
      summary: "Get related services of a selected service by ID"
      operationId: "getRelatedServicesByID"
      produces:
      - "application/json"
      parameters:
      - name: "id_service"
        in: "path"
        description: "ID of service to search for related"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ServiceItem"
        "400":
          description: "Invalid ID"
        "404":
          description: "Service not found"
      x-swagger-router-controller: "Service"
  /service/{id_service}/events:
    get:
      tags:
      - "service"
      summary: "Get events of a selected service by ID"
      operationId: "getServiceEventsByID"
      produces:
      - "application/json"
      parameters:
      - name: "id_service"
        in: "path"
        description: "ID of service to search for events"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/EventItem"
        "400":
          description: "Invalid ID"
        "404":
          description: "Service not found"
      x-swagger-router-controller: "Service"
  /service/{id_service}/people:
    get:
      tags:
      - "service"
      summary: "Get people of a selected service by ID"
      operationId: "getServicePeopleByID"
      produces:
      - "application/json"
      parameters:
      - name: "id_service"
        in: "path"
        description: "ID of service to search for people"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/PersonItem"
        "400":
          description: "Invalid ID"
        "404":
          description: "Service not found"
      x-swagger-router-controller: "Service"
  /service/{id_service}/articles:
    get:
      tags:
      - "service"
      summary: "Get articles of a selected service by ID"
      operationId: "getServiceArticlesByID"
      produces:
      - "application/json"
      parameters:
      - name: "id_service"
        in: "path"
        description: "ID of service to search for articles"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ArticleItem"
        "400":
          description: "Invalid ID"
        "404":
          description: "Service not found"
      x-swagger-router-controller: "Service"
...
  /people/items:
    get:
      tags:
      - "person"
      summary: "Get all people items"
      operationId: "getPeopleItemsAll"
      produces:
      - "application/json"
      parameters:
      - name: "page"
        in: "query"
        description: "People page number to get"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/PersonItem"
      x-swagger-router-controller: "Person"
  /person/{id_person}:
    get:
      tags:
      - "person"
      summary: "Get a person by ID"
      operationId: "getPersonByID"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id_person"
        in: "path"
        description: "ID of person to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Person"
        "400":
          description: "Invalid ID"
        "404":
          description: "Person not found"
      x-swagger-router-controller: "Person"
    delete:
      tags:
      - "person"
      summary: "Delete a person by ID"
      operationId: "deletePerson"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id_person"
        in: "path"
        description: "ID of person to be deleted"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid ID"
        "404":
          description: "Person not found"
      x-swagger-router-controller: "Person"
  /person:
    post:
      tags:
      - "person"
      summary: "Add a new person"
      operationId: "addPerson"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - in: "body"
        name: "person"
        description: "Person object to be added"
        required: true
        schema:
          $ref: "#/definitions/Person"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid input"
      x-swagger-router-controller: "Person"
    put:
      tags:
      - "person"
      summary: "Update a person"
      operationId: "updatePerson"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - in: "body"
        name: "person"
        description: "Person object to be updated"
        required: true
        schema:
          $ref: "#/definitions/Person"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid input"
        "404":
          description: "Person not found"
      x-swagger-router-controller: "Person"
  /person/{id_person}/services:
    get:
      tags:
      - "person"
      summary: "Get services of a selected person by ID"
      operationId: "getPersonServicesByID"
      produces:
      - "application/json"
      parameters:
      - name: "id_person"
        in: "path"
        description: "ID of person to search for services"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ServiceItem"
        "400":
          description: "Invalid ID"
        "404":
          description: "Person not found"
      x-swagger-router-controller: "Person"
  /person/{id_person}/events:
    get:
      tags:
      - "person"
      summary: "Get events of a selected person by ID"
      operationId: "getPersonEventsByID"
      produces:
      - "application/json"
      parameters:
      - name: "id_person"
        in: "path"
        description: "ID of person to search for events"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/EventItem"
        "400":
          description: "Invalid ID"
        "404":
          description: "Person not found"
      x-swagger-router-controller: "Person"
...
definitions:
  Service:
    type: "object"
    required:
    - "id_service"
    properties:
      id_service:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      presentation:
        type: "string"
      informations:
        type: "string"
      location:
        type: "string"
      photo_url:
        type: "string"
    xml:
      name: "Service"
    example:
      presentation: "presentation"
      name: "name"
      location: "location"
      photo_url: "photo_url"
      id_service: 0
      informations: "informations"
  ServicePage1:
    type: "object"
    required:
    - "id_service"
    properties:
      id_service:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      presentation:
        type: "string"
      photo_url:
        type: "string"
    xml:
      name: "ServicePage1"
    example:
      presentation: "presentation"
      name: "name"
      photo_url: "photo_url"
      id_service: 0
  ServicePage2:
    type: "object"
    required:
    - "id_service"
    properties:
      id_service:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      informations:
        type: "string"
    xml:
      name: "ServicePage2"
    example:
      name: "name"
      id_service: 0
      informations: "informations"
  ServiceItem:
    type: "object"
    required:
    - "id_service"
    properties:
      id_service:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      presentation:
        type: "string"
      photo_url:
        type: "string"
    xml:
      name: "ServiceItem"
    example:
      presentation: "presentation"
      name: "name"
      photo_url: "photo_url"
      id_service: 0
...
  EventItem:
    type: "object"
    required:
    - "id_event"
    properties:
      id_event:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      date:
        type: "string"
        format: "date"
      location:
        type: "string"
      photo_url:
        type: "string"
    xml:
      name: "EventItem"
    example:
      date: "2000-01-23"
      name: "name"
      location: "location"
      id_event: 0
      photo_url: "photo_url"
  Person:
    type: "object"
    required:
    - "id_person"
    properties:
      id_person:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      role:
        type: "string"
      email_address:
        type: "string"
        format: "email"
      phone_number:
        type: "string"
      description:
        type: "string"
      photo_url:
        type: "string"
      starting_date:
        type: "string"
        format: "date"
    xml:
      name: "Person"
    example:
      id_person: 0
      role: "role"
      email_address: "email_address"
      starting_date: "2000-01-23"
      name: "name"
      description: "description"
      phone_number: "phone_number"
      photo_url: "photo_url"
  PersonItem:
    type: "object"
    required:
    - "id_person"
    properties:
      id_person:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      photo_url:
        type: "string"
    xml:
      name: "PersonItem"
    example:
      id_person: 0
      name: "name"
      photo_url: "photo_url"
...
  ArticleItem:
    type: "object"
    required:
    - "id_article"
    properties:
      id_article:
        type: "integer"
        format: "int64"
      author:
        type: "string"
      publication_date:
        type: "string"
        format: "date"
      title:
        type: "string"
      photo_url:
        type: "string"
    xml:
      name: "ArticleItem"
    example:
      id_article: 0
      author: "author"
      publication_date: "2000-01-23"
      photo_url: "photo_url"
      title: "title"
...